parseInt to parseFloat

+++ assets/js/app.js
@@ -1,8 +1,8 @@
 //calculate inputs to determine charge amount
 function calculate() {
-	var hours = parseInt(document.getElementById("hours-input").value, 10);
-	var rate = parseInt(document.getElementById("rate-input").value, 10);
-	var supplies = parseInt(document.getElementById("cost-input"). value, 10);
+	var hours = parseFloat(document.getElementById("hours-input").value, 10);
+	var rate = parseFloat(document.getElementById("rate-input").value, 10);
+	var supplies = parseFloat(document.getElementById("cost-input"). value, 10);
 	var total = hours * rate + supplies;
 		document.getElementById("charge").innerHTML = total;
 };
